pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "content-automation"
version = "0.1.0"
description = "Local-first social media publishing application."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14,<2.0",
    "fastapi>=0.115,<1.0",
    "jinja2>=3.1,<4.0",
    "pillow>=11.0,<12.0",
    "pydantic-settings>=2.8,<3.0",
    "python-dotenv>=1.0,<2.0",
    "python-multipart>=0.0.9,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "uvicorn[standard]>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.0,<26.0",
    "httpx>=0.28,<1.0",
    "pytest>=8.0,<9.0",
    "ruff>=0.11,<1.0",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I"]

[tool.ruff.lint.isort]
known-first-party = ["app", "tests"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
app = ["templates/**/*.html", "static/**/*.css"]
